chore: expose datanode as subcommand of vega
